Исправит ошибки е2е тестов
