+++ src/components/review-item/review-item.tsx
@@ -14,38 +14,34 @@ function ReviewItem({reviewItem}: ReviewProps): JSX.Element {
   }).format(new Date(date.split('T')[0]));
 
   return (
-    <ul className="reviews__list">
-      <li className="reviews__item">
-        <div className="reviews__user user">
-          {avatarUrl && (
-            <div className="reviews__avatar-wrapper user__avatar-wrapper">
-              <img
-                className="reviews__avatar user__avatar"
-                src={avatarUrl}
-                width={54}
-                height={54}
-                alt="Reviews avatar"
-              />
-            </div>
-          )}
-          <span className="reviews__user-name">{name}</span>
+    <li className="reviews__item">
+      <div className="reviews__user user">
+        <div className="reviews__avatar-wrapper user__avatar-wrapper">
+          <img
+            className="reviews__avatar user__avatar"
+            src={avatarUrl}
+            width={54}
+            height={54}
+            alt="Reviews avatar"
+          />
         </div>
-        <div className="reviews__info">
-          <div className="reviews__rating rating">
-            <div className="reviews__stars rating__stars">
-              <span style={{ width: `${handleStars(rating)}` }} />
-              <span className="visually-hidden">Rating</span>
-            </div>
+        <span className="reviews__user-name">{name}</span>
+      </div>
+      <div className="reviews__info">
+        <div className="reviews__rating rating">
+          <div className="reviews__stars rating__stars">
+            <span style={{ width: `${handleStars(rating)}` }} />
+            <span className="visually-hidden">Rating</span>
           </div>
-          <p className="reviews__text">
-            {comment}
-          </p>
-          <time className="reviews__time" dateTime={date.split('T')[0]}>
-            {dueDate}
-          </time>
         </div>
-      </li>
-    </ul>
+        <p className="reviews__text">
+          {comment}
+        </p>
+        <time className="reviews__time" dateTime={date.split('T')[0]}>
+          {dueDate}
+        </time>
+      </div>
+    </li>
   );
 }
